### 2020-08-17 #2578 @whedon
[REVIEW]: Off-resonance CorrecTion OPen soUrce Software (OCTOPUS)
**Submitting author:** @imr-framework (<a href="http://orcid.org/0000-0002-1141-2049">Marina Manso Jimeno</a>)
**Repository:** <a href="https://github.com/imr-framework/OCTOPUS" target ="_blank">https://github.com/imr-framework/OCTOPUS</a>
**Version:** v1.0.0
