### 2020-08-25 #2578 @puolival
@jni I am almost finished with the review (1–2 more days), so I thought I would send all comments at once.
