### 2020-08-28 #2578 @jni
@puolival

> this requirement is only met for @Mmj94

This is a collective requirement for the paper as a whole, not meant to be interpreted as needing 3 months of work *per author*. So I think this means you agree that the requirement is met?
