### 2020-08-28 #2578 @puolival
@jni Sorry, I thought it was per person. I do agree that the requirement is met collectively. I have now updated the checklist.
