### 2020-09-01 imr-framework/OCTOPUS#1 @emilljungberg
JOSS Review: read_dicom image flip
Hi, I'm reviewing your submission of `OCTOPUS` to JOSS and when I looked through the code I found some small bugs I wanted to bring your attention to.

In the `read_dicom` the last line before the return statement righ-left flips the image with

```python
vol = np.fliplr(vol)
```

I understand that this line might solve a specific orientation problem for your application but as `read_dicom` appears to be a general DICOM reader in your project, it can cause quite serious problems for users who do not know that their data is being flipped.

Re: [openjournals/joss-reviews/#2578](https://github.com/openjournals/joss-reviews/issues/2578)
