### 2020-09-01 imr-framework/OCTOPUS#2 @emilljungberg
JOSS Review: sum-of-squares calculation
In the `main_ORC.py` example code there is a sum of squares coil combination call at the very end. There seems to be a small mistake in this calculation. Currently, the line reads
```python
sos_CPR = np.sum(np.abs(CPR_result), -1)
```

It should probably be:
```python
sos_CPR = np.sqrt(np.sum(np.abs(CPR_result)**2, -1))
```

Re: [openjournals/joss-reviews/#2578](https://github.com/openjournals/joss-reviews/issues/2578)
