### 2020-09-01 imr-framework/OCTOPUS#3 @emilljungberg
JOSS Review: Reproduce figures from paper
I ran the Cartesian demo (`numsim_cartesian.py`) and Non-Cartesian demo (`numsim_spiral.py`) from the examples folder to generate the figure in the manuscript but I seems like the figures I get are not the same as in the manuscript. (ignore the difference in the labels, that's just a viewer/font-size issue)

The **Cartesian result** appear right-left flipped compared to the manuscript
![cartesian_results](https://user-images.githubusercontent.com/3906827/91837008-df395400-ec4b-11ea-8bde-640e8358a344.png)

and the **Spiral results** appear to have much less distortions than in the manuscript
![spiral_results](https://user-images.githubusercontent.com/3906827/91837080-f6784180-ec4b-11ea-98d2-783cf40f35e9.png)

Related to this, why is the frequency offset range different for the Cartesian and spiral simulations? +/- 4800 Hz seems like a very large off-resonance range. B0 maps would typically span the range of about +/- 100Hz, making the spiral simulation seems to be more in the realistic range.

Re: [openjournals/joss-reviews/#2578](https://github.com/openjournals/joss-reviews/issues/2578)
