### 2020-09-01 imr-framework/OCTOPUS#4 @emilljungberg
JOSS Review: Data Structures documentation
The documentation provided for this software is very nicely outlined but it would be great to have some more information about the data structures used:

**k-space and image data**: How should these be structured? Cartesian data is quite straight forward but when it comes to non-cartesian and multi-channel/coil data it is not obvious how to structure it. A short explanation for how to structure these data would be helpful.

**non-cartesian trajectories**: Similar to the k-space and image data, description of the layout of the trajectories and which order the dimensions should be in would be very useful.

**Non-Cartesian parameter dictionary**: The NUFFT object requires a sequence parameter object that appears to be related to the trajectory. Could you please document how this dictionary should be structured?

Re: [openjournals/joss-reviews/#2578](https://github.com/openjournals/joss-reviews/issues/2578)
