### 2020-09-01 imr-framework/OCTOPUS#5 @emilljungberg
JOSS Review: Package structure
Packages and modules are typically named with all lowercase letters (e.g. `recon` instead of `Recon` and `fieldmap` instead of `Fieldmap`). [PEP 8 - package and module names](https://www.python.org/dev/peps/pep-0008/#package-and-module-names)

To clarify and generalize the use of the other submodules, some suggestions:
  - ~~`OCTOPUS.utils.plot_results`~~ -> `OCTOPUS.utils.plotting`
  - There are two modules for data reading `OCTOPUS.Recon.read_dicom` and `OCTOPUS.utils.get_data_from_file`, each containing just a single function. Consider combining to a single module with naming such as `dataio`.
  - In the `OCTOPUS.Fieldmap` sub-package, the `fieldmap_gen` module could be renamed to `simulate` as it is used to simulate field maps and would produce a more natural naming (`OCTOPUS.fieldmap.simulate`). Since the sub-package is called `fieldmap`, that word does not have to be repeated in the module name.
  - The `phase_unwrap_prep` function does not appear to be used and is specific to use with `fsl`. I'd consider renaming the module to something like `unwrap` and have functions in there that are used for very specific applications like the FSL prep, or moving it to `utils`.

Re: [openjournals/joss-reviews/#2578](https://github.com/openjournals/joss-reviews/issues/2578)
