### 2020-09-07 #2578 @Mmj94
@jni @emilljungberg @puolival Thank you so much for your comments and reviews.
Sorry for the delay in my response, I just came back from vacation but I'll definitely start working on all the issues asap.
