### 2020-09-08 #2578 @emilljungberg
# Review Summary
The `OCTOPUS` software package consists of python implementations of three different off-resonance correction methods for MRI reconstruction. The package was designed with the intention of making implementation of these methods easier to use, with the motivation that current available software packages, typically in MATLAB, are not flexible enough to accommodate different acquisition strategies. I think the intention behind this work is really good but both the manuscript and the code need improvement before it can fill the gap that the authors have identified.

The python code is well structured into sub-packages and modules with very clear documentation for each function. Many functions have been developed to simplify the interface for the end user, such as `im2ksp` and `ksp2im`, as well as generalized data input functions `get_data_from_file`. However, in doing so, I think the authors end up in the same trap that they tried to avoid, namely not making their tool flexible to all types of data and acquisition strategies. These concerns are highlighted more specifically later in the review.

The demos provided in the `OCTOPUS` repository clearly demonstrates that the tool is able to correct off-resonance artefacts with the three methods provided. However, it is not clear to me if `OCTOPUS` is intended to be an end-to-end reconstruction package which includes off-resonance correction, or as a tool for use together with other reconstruction toolboxes. As a complete reconstruction package, `OCTOPUS` only implements the off-resonance correction and wrappers for 2D-FFT and NUFFT and lacks the advanced reconstruction methods used today such as parallel imaging and compressed sensing.

In the sections I provide comments on the manuscript, the code and some concluding recommendations.

## Comments on the manuscript

### Summary
- The summary section could be expanded to be more specific as to which methods it implements and to which extent it is _flexible to the acquisition scheme_. I assume that it refers to handling both Cartesian and non-cartesian data, but that should be specified, and it should also be clarified if it can handle 2D and/or 3D data.
- I don't think the web browser feature needs to be listed here, and its utility overall should be highlighted better later in the manuscript.

### Statement of need
- This section provides an overview why off-resonance correction is needed and previous methods. The third paragraph outlines existing software packages for off-resonance correction with the initial statement that they have __limited off-resonance correction capabilities__. I don't think this statement is justified. For instance, the SPIRiT paper outlines how the method work for both Cartesian and non-Cartesian acquisitions. Similarily Sutton et al. demonstrated their method to work for non-Cartesian data, which means it would work for Cartesian as well. The next paragraph follow the same path, unfairly critisising the exisiting tools for being limited without specifying in which way they are limited. The argument that MATLAB based methods are restricting portability, accessbility and customisation of the code is also not justified in my opinion. The reference to Ravi et al. does not support this argument either as their paper discusses a specific implementation of a GUI for development of MR pulse sequences. A different reference would be needed to support the statement that MATLAB is a limiting factor in research workflows. One could almost make the opposite argument, many reconstruction toolboxes already developed for MATLAB (such as the SPIRiT package, BART, MIRT, GE Orchestra, Gadgetron, GPUnufft, MRiLAB) and therefore, a MATLAB interfacing would allow for better integration with other tools. Although, many of these do also support Python now.
- The use of a Google colab notebooks for compatible implementations is unclear to me, how would this simplify a reconstruction pipeline? This comes up in the last sentence of the last paragraph again where `OCTOPUS` is refered to as a **zero-footprint** software since it can be run in a web-browser. I'm not familiar with the term **zero-footprint software** and again it is not clear how a web-browser implementation helps in a MR reconstruction workflow which requires reading and writing of raw data, tools which are typically proprietary and needs to be run on a local machine.
- The key sentence in this section is in the last paragraph which lists the three methods that are implemented in this toolbox. It would be good to highlight these better with a bullet point list for instance, and include the references for each method.

#### Specific comments on writing
> Off-resonance is an MR artifact ~~whose source can vary from~~ **which can be produced by for instance** field inhomogeneities **,** ~~and to~~ differences in tissue susceptibilities and chemical shift

> These phenomena can cause the phase of off-resonant **spins** ~~(not at the resonant or Larmor frequency) spins~~ to accumulate along the read-out direction.

Users of this toolbox will be familiar with the term __off-resonant spins__.

### Short demo
- It would be good to include the commands necessary to run the demos here.

### Specific points to address
> __Summary:__ Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

It would be good to highlight the three specific methods that this package includes and what the limitations are such as 2D/3D images and Cartesian/non-cartesian trajectories.

> __A statement of need:__ Do the authors clearly state what problems the software is designed to solve and who the target audience is?

It is clear that the software corrects off-resonance artefacts but it could be clarified who the software is for and what the typical use of the software would be. Is it intended to mainly operate on reconstructed images or on raw k-space data?

> __State of the field:__ Do the authors describe how this software compares to other commonly-used packages?

I think this needs to be clarified in the manuscript. It is stated that available packages are limited and highly specific but without specifying in which way they are specific and how `OCTOPUS` fills this gap. Considering that off-resonance correction is intended to remove an image artefact, and that it is a computationally heavy problem to solve, comparison to existing packages in terms of image quality and processing time would be useful to evaluate the package.

> __Quality of writing:__ Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

The language and writing style in the manuscript is generally of good quality. As mentioned by @puolival, there are a few typos that needs to be checked.

> __References:__ Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper citation syntax?

It would be good to extend the reference list with URLs to the available code for the other methods listed in the statement of needs section.

## Specific comments on the code
- The code is very well organised as a Python package and subpackages, but it could potentially be simplified. A few of the subpackages include modules with only a single function. It would probably be possible to simplify the whole package to have a main package with several modules instead of subpackage (i.e. folders).
- Consider using a numerical Shepp-Logan phantom, such as the one in `skimage.data` to reduce the data size in the repository. It would also enable more flexible demos.
- Much of the code seems to be limited to 2D applications but that isn't clearly stated anywhere.
- The `fmap_recon` appears to only be available for Cartesian data.
- The `SSIM_calc` and `pSNR_calc` functions use the data range of the second image to set the window range for the calculation of SSIM and pSNR. If the two input images have very different data range this could skew the results. The avoid this, it could be better to use the `scikit` function directly instead, especially given that `SSIM_calc` and `pSNR_calc` only wraps this function and simplifies one parameter call. This eliminate any potential errors.
- Consider adding in references to the `ORC` methods in the `ORC` module (i.e. for CPR, fs-CPR, and MFI).

## General Checks
> **Substantial scholarly effort**: Does this submission meet the scope eligibility described in the JOSS guidelines

Based on an assessment of the code for this package, I do not think the current code base satisfies the JOSS scope eligibility criteria. The core of this package is in the `ORC` module which implements the three off-resonance correction modules in less than 500 lines of code, much of which is doc strings and repeated lines of code for data input. Much of the repository consists of code to demonstrate the use of the toolbox or wrappers to simplify function calls such as FFT or image metric calculations from other packages. Data input/output methods that would be required to make this a powerful package are too simple to be useful, or too specific to the authors use case to be generalisable.

However, I think there is potential for this package to be really useful but it would, in my view, require either extending the package or pivoting in one of a few different ways, which I have outlined in the last section (Suggestions for moving forward)

## Functionality
> **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

There are no comments on the performance of the code which would have been good to include. A comparison between the different methods would be really useful for users, given that the CPR method is considerable slower than fs-CPR and MFI (~20min for a single slice and coil, compared to about 20s for the two other methods).

Again regarding the use of this tool for image or k-space data, would it work with parallel imaging undersampled data where a direct FFT/NUFFT of the data would result in folding artefacts? I think this needs to be clarified as almost all acquisitions use some type of image acceleration.

## Documentation
> **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

There is a very nice read the docs API documentation provided for the repository. It would however be good to be more specific regarding the different data structures used. For instance:
  - How should image and raw data (Cartesian and non-cartesian) be structured?
  - How should the non-Cartesian trajectory be arranged?
  - The sequence parameters dictionary used for non-cartesian acquisitions should be described.

> **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

The authors have developed unit tests for the core functionality but it could be expanded. In particular it would be good to test the ORC for input data size (1D/2D/3D/4D) and for input data type (magnitude/complex)

## Suggestions for moving forward
To make the best use and gain the widest traction of the effort put into this software package, I think it would be good to distinguish between correction of image data and k-space data, as the latter requires additional advanced reconstruction methods.

For an image correction method, the end user would need a robust data input and output method, typically reading either dicom or nifti, together with a field map input or data from which a field map can be reconstructed. For this to be really useful, it would be good to have a python script which can take the input data as commandline argument, choose the correction method, and return a corrected image. That would take this tool to a useful research tool that people could use and modify to their preferences.

For a k-space based correction method on the other hand, the user would need to know how to integrate this tool with other reconstruction toolboxes to produce their final image. A root sum of squares coil combination is rarely sufficient these days and so there needs to be a clear description of how to integrate this method with other reconstruction toolboxes, such as the `sigpy` framework in python (which I have no affiliation to). To enable this, there would need to be a bit more documentation of the data structures. Also, for processing of raw data, would the methods work with parallel imaging data where a direct FFT of the data results in folding artefacts? For a reconstruction framework, this is a crucial part as almost all acquisitions today use some kind of acceleration.

Overall, the functionality of the methods in `OCTOPUS` is clear from the examples, but the functionality in terms of compatible input data and how to integrate it into a research framework is not as clear to me. With additional documentation and some development of the code I think this can be resolved.
