### 2020-09-09 2df1f42
Remove rotation from read_dicom
