### 2020-09-10 23d9eb8
Fix sum of squares calculation in ORC_main.py
