### 2020-09-14 9291b18
Rename OCTOPUS.utils.plot_results to plotting (Issue #5.2)
