+++ OCTOPUS/Examples/numsim_spiral.py
@@ -11,7 +11,7 @@ import matplotlib.pyplot as plt
 
 import OCTOPUS.ORC as ORC
 import OCTOPUS.Fieldmap.fieldmap_gen as fieldmap_gen
-from OCTOPUS.utils.plot_results import plot_correction_results
+from OCTOPUS.utils.plotting import plot_correction_results
 from OCTOPUS.utils.metrics import create_table
 ##
 # Original image: Shep-Logan Phantom
