### 2020-09-14 db32892
Recon sub-package to lowercase (Issue #5.1)
