+++ OCTOPUS/Examples/display_panel_images.py
@@ -6,7 +6,7 @@ Last updated: 07/08/2020
 '''
 import numpy as np
 import configparser
-from OCTOPUS.Recon.read_dicom import read_dicom
+from OCTOPUS.recon.read_dicom import read_dicom
 from OCTOPUS.utils.plotting import plot_correction_results
 from OCTOPUS.utils.metrics import create_table
 
