+++ OCTOPUS/Fieldmap/prepare_fieldmap_unwrap.py
@@ -8,10 +8,12 @@ import os
 import scipy.io as sio
 import numpy as np
 import nibabel as nib
+import matplotlib.pyplot as plt
 
 from pydicom import dcmread
 
-from OCTOPUS.Recon.rawdata_recon import mask_by_threshold
+from OCTOPUS.utils.get_data_from_file import get_data_from_file
+from OCTOPUS.recon.rawdata_recon import mask_by_threshold
 
 def phase_unwrap_prep(data_path_raw, data_path_dicom, dst_folder, dTE):
     '''
@@ -29,12 +31,14 @@ def phase_unwrap_prep(data_path_raw, data_path_dicom, dst_folder, dTE):
     dTE : float
         Difference in TE between the two echoes in seconds
     '''
-    if data_path_raw[-3:] == 'mat':
+    b0_map = get_data_from_file(data_path_raw)
+    '''if data_path_raw[-3:] == 'mat':
         b0_map = sio.loadmat(data_path_raw)['b0_map'] # field map raw data
     elif data_path_raw[-3:] == 'npy':
         b0_map = np.load(data_path_raw)
     else:
-        raise ValueError('File format not supported, please input a .mat or .npy file')
+        raise ValueError('File format not supported, please input a .mat or .npy file')'''
+
 
     ##
     # Acq parameters
@@ -47,7 +51,11 @@ def phase_unwrap_prep(data_path_raw, data_path_dicom, dst_folder, dTE):
         b0_map = b0_map.reshape(b0_map.shape[0], N, 1, 2, Nchannels)
     else:
 
-        Nslices = b0_map.shape[-2]
+        Nslices = b0_map.shape[-3]
+        sl_order = np.zeros((Nslices))
+        sl_order[range(0,Nslices,2)] = range(int(Nslices/2), Nslices)
+        sl_order[range(1,Nslices,2)] = range(0, int(Nslices/2))
+        sl_order = sl_order.astype(int)
 
     ##
     # FT to get the echo complex images
@@ -65,18 +73,37 @@ def phase_unwrap_prep(data_path_raw, data_path_dicom, dst_folder, dTE):
     echo1 = echo1[oversamp_factor:-oversamp_factor, :, :, :]
     echo2 = echo2[oversamp_factor:-oversamp_factor, :, :, :]
 
+    echo1 = echo1[:,:,sl_order,:]
     # Magnitude image with object 'brain' extracted
-    mag_im = np.sum(np.abs(echo1), 3)
+    mag_im = np.sum(np.abs(echo1), -1)
     mag_im = np.divide(mag_im, np.max(mag_im))
+
+
+    im = np.zeros((128, 128 * 20))
+    for i in range(20):
+        im[:, int(128 * i):int(128 * (i + 1))] = mag_im[:, :, i]
+    plt.imshow(im)
+    plt.show()
+
+
     brain_mask = mask_by_threshold(mag_im)
-    brain_extracted = np.squeeze(mag_im) * brain_mask
+    brain_extracted = np.rot90(np.squeeze(mag_im) * brain_mask)
+
     img = nib.Nifti1Image(brain_extracted, np.eye(4))
     nib.save(img, os.path.join(dst_folder, 'mag_vol_extracted.nii.gz'))
 
     # Phase difference image from DICOM data
-    data = dcmread(data_path_dicom)
-    # Get the image data
-    vol = data.pixel_array
+    if data_path_dicom[-4] != '.dcm' or data_path_dicom[-4:] != '.IMA':
+        # multi-slice
+        files_in_folder = os.listdir(data_path_dicom)
+        vol = np.zeros((N, N, Nslices))
[…]
