### 2020-09-14 3fb7f99
Recon sub-package to lowercase (Issue #5.1)
