### 2020-09-14 1b764b2
Recon sub-package to lowercase (Issue #5.1)
