### 2020-09-14 60772dc
Recon sub-package to lowercase (Issue #5.1)
