### 2020-09-14 0660a1a
Fieldmap sub-package to lowercase (Issue #5.1)
