### 2020-09-14 f2b7c3f
Fieldmap sub-package to lowercase (Issue #5.1)
