### 2020-09-14 89d53c8
fieldmap.fieldmap_gen to fieldmap.simulate (Issue #5.4)
