### 2020-09-14 c0d1d31
combine read_dicom and get_data_from_file into dataio (Issue #5.3)
