+++ OCTOPUS/Examples/ORC_main.py
@@ -13,7 +13,7 @@ import math
 import time
 import os
 
-from OCTOPUS.utils.get_data_from_file import get_data_from_file
+from OCTOPUS.utils.dataio import get_data_from_file
 from OCTOPUS.recon.rawdata_recon import fmap_recon
 from OCTOPUS.recon.rawdata_recon import spiral_recon
 from OCTOPUS.utils.plotting import plot_correction_results
