+++ OCTOPUS/recon/rawdata_recon.py
@@ -10,7 +10,7 @@ import matplotlib.pyplot as plt
 import numpy.fft as fft
 import math
 
-from OCTOPUS.utils.get_data_from_file import get_data_from_file
+from OCTOPUS.utils.dataio import get_data_from_file
 from OCTOPUS.recon.imtransforms import nufft_init, ksp2im
 
 def mask_by_threshold(im):
