### 2020-09-14 85e7697
Document data structure for correction input (Issue #4)
