### 2020-09-15 imr-framework/OCTOPUS#1 @Mmj94
Thank you for pointing this bug out.  Commit 2df1f4207f2fc5caa8bcec3b19fd20f68d54c143 should fix it.
Please let me know if I can close the issue.
