### 2020-09-15 imr-framework/OCTOPUS#2 @Mmj94
You're right! Same mistake applies to `sos_fsCPR` and `sos_MFI`. Thank you for noticing!
Commit 23d9eb8e01c2be62698e5b0e379b0ef583f48ef4 should fix this problem, let me know and I'll close this issue.
