### 2020-09-15 imr-framework/OCTOPUS#3 @Mmj94
1. You're right, it is not possible to exactly reproduce the figures in the manuscript. This is because the field map simulation used for both the cartesian and non-cartesian demos is random. However, I believe I saved the field maps for the figures shown in the manuscript (There is some commented code related to this in both scripts titled "For reproducibility").  Let me know if you would like to exactly reproduce those figures and I'll upload the files to the repository.

2. Regarding your second point: The off-resonance offset is very large for the cartesian case just because we are exploring reconstruction techniques in highly inhomogeneous fields for a different project. Apart from that, the cartesian trajectory ultra-short readout time makes it less prone to off-resonance artifacts, therefore, needing a much larger frequency range compared to the spiral case to show equivalent image distortion.

Please let me know if these comments are helpful and if I can close this issue.
