### 2020-09-15 imr-framework/OCTOPUS#5 @Mmj94
Thank you for your helpful comments!

Please let me know if I can now close the issue.
