### 2020-09-15 imr-framework/OCTOPUS#3 @emilljungberg
Ah right I see I didn't realise that. But looking through the code now that makes sense. It could be an idea to use a seeded random number generator in your example script which would ensure that your samples are pseudo random, such that you get the same random numbers every time. It would allow the user to make sure they have everything working correctly if they can reproduce the same images as in the paper.

Example usage in `numpy` discussed in this thread for instance:
https://stackoverflow.com/questions/22994423/difference-between-np-random-seed-and-np-random-randomstate
