### 2020-09-15 imr-framework/OCTOPUS#4 @Mmj94
Please have a look at the commit above, I added some documentation to ORC.py regarding your suggestions.
