### 2020-09-15 8711fea
Add random seed for field map reproducibility
