### 2020-09-23 imr-framework/OCTOPUS#3 @emilljungberg
Would you be able to regenerate the figures on your end so that I can reproduce them with this fix? Thanks!
