### 2020-10-01 5c1a01f
Minor code fixes + command line prog.py
