+++ OCTOPUS/Examples/numsim_spiral.py
@@ -8,6 +8,8 @@ import numpy as np
 import math
 import cv2
 import matplotlib.pyplot as plt
+from skimage.data import shepp_logan_phantom
+from skimage.transform import resize
 
 import OCTOPUS.ORC as ORC
 import OCTOPUS.fieldmap.simulate as fieldmap_sim
@@ -17,9 +19,8 @@ from OCTOPUS.utils.metrics import create_table
 # Original image: Shep-Logan Phantom
 ##
 def numsim_spiral():
-    ph = np.load('sample_data/slph_im.npy').astype(complex) # Shep-Logan Phantom
-    ph = (ph - np.min(ph)) / (np.max(ph)-np.min(ph)) # Normalization
-    N = ph.shape[0]
+    N = 192  # ph.shape[0]
+    ph = resize(shepp_logan_phantom(), (N, N)).astype(complex)
     plt.imshow(np.abs(ph), cmap='gray')
     plt.title('Original phantom')
     plt.axis('off')
