### 2020-10-08 c26d3e9
Use numerical shepp-logan phantom for simulations
