### 2020-10-08 07dac2a
Comment existent tests and add more
