+++ tests/utests_ORC_methods.py
@@ -6,7 +6,7 @@ import numpy as np
 
 import OCTOPUS.ORC as ORC
 
-from OCTOPUS.Fieldmap.fieldmap_gen import fieldmap_bin
+from OCTOPUS.fieldmap.simulate import fieldmap_bin
 
 # Inputs
 raw_data = sio.loadmat('test_data/acrph_noncart_ksp.mat')['dat']
@@ -22,25 +22,31 @@ acq_params = {'Npoints': ktraj.shape[0], 'Nshots': ktraj.shape[1], 'N': field_ma
 
 class TestfscPR(unittest.TestCase):
     def test_fsCPR_given_rawData(self):
+        # Test correction given raw data
         or_corr_im = ORC.fs_CPR(dataIn=raw_data[:,:,0], dataInType='raw', kt=ktraj, df=field_map, Lx=1, nonCart=1, params=acq_params)
         self.assertEqual(or_corr_im.shape, field_map.shape[:-1]) # Dimensions agree
 
     def test_fsCPR_given_im(self):
+        # Test correction given image data
         or_corr_im = ORC.fs_CPR(dataIn=raw_data[:,:,0],dataInType='raw',kt=ktraj,df=field_map,Lx=1, nonCart=1, params=acq_params)
-        or_corr_im2 = ORC.fs_CPR(dataIn=or_corr_im,dataInType='im',kt=ktraj,df=field_map,Lx=1,nonCart=1,params=acq_params)
-        self.assertEqual(or_corr_im2.shape, field_map.shape[:-1])
+        or_corr_im2 = ORC.fs_CPR(dataIn=or_corr_im,dataInType='im',kt=ktraj,df=np.squeeze(field_map),Lx=1,nonCart=1,params=acq_params)
+        self.assertEqual(or_corr_im2.shape, field_map.shape[:-1]) # Dimensions match
         self.assertEqual(or_corr_im2.all(), or_corr_im.all())
 
     def test_fsCPR_wrongtype(self):
+        # Error when dataIn is raw data but dataInType is 'im'
         with self.assertRaises(ValueError): ORC.fs_CPR(dataIn=raw_data[:, :, 0], dataInType='im', kt=ktraj, df=field_map, Lx=1, nonCart=1, params=acq_params)
         or_corr_im = ORC.fs_CPR(dataIn=raw_data[:, :, 0], dataInType='raw', kt=ktraj, df=field_map, Lx=1,
                                 nonCart=1, params=acq_params)
+        # Error when dataIn is image data but dataInType is 'raw'
         with self.assertRaises(ValueError): ORC.fs_CPR(dataIn=or_corr_im, dataInType='raw', kt=ktraj,
                                                        df=field_map, Lx=1, nonCart=1, params=acq_params)
+        # Error when dataInType is other than 'raw' or 'im'
         with self.assertRaises(ValueError): ORC.fs_CPR(dataIn=raw_data[:, :, 0], dataInType='other', kt=ktraj,
                                                        df=field_map, Lx=1, nonCart=1, params=acq_params)
 
     def test_find_nearest(self):
+        # test that find_nearest works for positive and negative values
         array = np.linspace(-10, 10, 21)
         neg_val = -5
         pos_val = 5
@@ -51,11 +57,13 @@ class TestfscPR(unittest.TestCase):
 
 class TestMFI(unittest.TestCase):
     def test_MFI_given_rawData(self):
+        # Test correction given raw data
         or_corr_im = ORC.MFI(dataIn=raw_data[:, :, 0], dataInType='raw', kt=ktraj, df=field_map[:,:,0], Lx=1,
                                 nonCart=1, params=acq_params)
[…]
