### 2020-10-08 ec05802
Delete pSNR and SSIM wrapper functions
