### 2020-10-08 15bafdb
Dicom reader to support dicom folders and dicom files
