### 2020-10-08 3e9ca46
Add some dimension documentation + check functions
