+++ OCTOPUS/ORC.py
@@ -2,7 +2,7 @@
 '''
 Methods for ORC (Off-Resonance Correction) and off-resonance simulation
 \nAuthor: Marina Manso Jimeno
-\nLast updated: 07/14/2020
+\nLast updated: 10/07/2020
 
 \n Some comments on data input and output:
 \n At the current stage, the methods apply a 2D reconstruction only. Therefore, data input and output will be two-dimensional.
@@ -31,6 +31,81 @@ import numpy as np
 from math import ceil, pi
 
 from OCTOPUS.recon.imtransforms import ksp2im, im2ksp, nufft_init
+
+
+def check_inputs_cartesian(dataInShape, dataInType, ktShape, dfShape):
+    '''Check dimensions of the inputs
+
+    Parameters
+    ----------
+    dataInShape :  tuple
+        Shape of input data. Must be [NxN] for image data and [NlinesxNcolumns] for raw data.
+    dataInType : str
+        Type of data: im or raw
+    ktShape : tuple
+        Shape of k-space trajectory. Must be [NlinesxNcolumns].
+    dfShape : tuple
+        Shape of field map. Must be [NxN] and match the image matrix size.
+    '''
+    if dataInType == 'im':
+        if dataInShape[0] != dataInShape[1]:
+            raise ValueError('Data dimensions do not agree with expected image dimensions (NxN)')
+        if dataInShape != dfShape:
+            raise ValueError('The image and field map dimensions do not match')
+    elif dataInType == 'raw':
+        if dataInShape != ktShape:
+            raise ValueError('The raw data and k-space trajectory dimensions do not match')
+    else:
+        raise ValueError('This type of input data is not supported. Please select im or raw')
+
+
+def check_inputs_noncartesian(dataInShape, dataInType, ktShape, dfShape, params):
+    '''Check dimensions of the inputs
+
+    Parameters
+    ----------
+    dataInShape :  tuple
+        Shape of input data. Must be [NxN] for image data and [NpointsxNshots] for raw data.
+    dataInType : str
+        Type of data: im or raw
+    ktShape : tuple
+        Shape of k-space trajectory. Must be [NpointsxNshots].
+    dfShape : tuple
+        Shape of field map. Must be [NxN] and match the image matrix size.
+    params : dict
+        Sequence parameters. Required elements are Npoints, Nshots, and N.
+    '''
+    if dataInType == 'im':
+        if dataInShape[0] != dataInShape[1]:
+            raise ValueError('Data dimensions do not agree with expected image dimensions (NxN)')
+        if dataInShape != dfShape:
+            raise ValueError('The image and field map dimensions do not match')
+
+    elif dataInType == 'raw':
+        if dataInShape != ktShape:
+            raise ValueError('The raw data and k-space trajectory dimensions do not match')
+
+    else:
+        raise ValueError('This type of input data is not supported. Please select im or raw')
+
+    if 'Npoints' not in params:
+        raise ValueError('The number of acquisition points is missing')
+    if 'Nshots' not in params:
+        raise ValueError('The number of shots is missing')
+    if 'N' not in params:
+        raise ValueError('The matrix size N is missing')
+
+    if params['N'] != dfShape[0]:
+        raise ValueError('The image or field map dimensions and value specified for N do not match')
+
+    if params['Npoints'] != ktShape[0]:
[…]
