### 2020-10-08 d5a6f4b
Command line implementation
