+++ OCTOPUS/prog.py
@@ -3,6 +3,7 @@ import sys
 import os
 import numpy as np
 from tqdm import tqdm
+import math
 
 sys.path.insert(0,r'C:\Users\marin\Documents\GitHubRepositories\OCTOPUS')
 
@@ -11,14 +12,16 @@ import OCTOPUS.ORC as ORC
 
 
 parser = argparse.ArgumentParser()
-parser.add_argument("path", help="echo the string you use here")
-parser.add_argument("datain", help="echo the string you use here")
-parser.add_argument("ktraj", help="echo the string you use here")
-parser.add_argument("fmap", help="echo the string you use here")
-parser.add_argument("ORCmethod", choices=['CPR', 'fsCPR', 'MFI'], help="echo the string you use here")
-parser.add_argument('--Lx', default=2, dest='Lx')
-parser.add_argument('--grad_raster', default=10e-6, dest='grad_raster')
-parser.add_argument('--TE', default=0, dest='TE')
+parser.add_argument("path", help="Path containing the input data")
+parser.add_argument("datain", help="Corrupted data with dimensions [Npoints OR N(Cartesian), Nshots OR N(Cartesian), Nslices, Nchannels]")
+parser.add_argument("ktraj", help="K-space trajectory: [Npoints OR N(Cartesian), Nshots OR N(Cartesian)]")
+parser.add_argument("fmap", help="Field map in rad/s")
+parser.add_argument("ORCmethod", choices=['CPR', 'fsCPR', 'MFI'], help="Correction method")
+parser.add_argument("--cart", default=0, choices=[0, 1], dest="cart_opt", help="0: non-cartesian data, 1: cartesian data")
+parser.add_argument('--Lx', default=2, dest='Lx', help="L(frequency bins) factor with respect to minimum L")
+parser.add_argument('--grad_raster', default=10e-6, dest='grad_raster', help="Gradient raster time")
+parser.add_argument('--TE', default=0, dest='TE', help="Echo time")
+parser.add_argument('--dcf', default=None, dest='dcf', help="Density compensation factor for non-cartesian trajectories")
 
 #li = [r'C:\Users\marin\Documents\PhD\B0inhomogeneity\Data\20200917\shimmed', 'rawdata_spiral.mat', 'ktraj.npy', 'fieldmap_unwrapped.nii.gz', 'MFI']
 
@@ -28,7 +31,7 @@ args = parser.parse_args()
 
 data_in = get_data_from_file(os.path.join(args.path, args.datain))
 ktraj = get_data_from_file(os.path.join(args.path, args.ktraj))
-fmap = get_data_from_file(os.path.join(args.path, args.fmap))
+fmap = get_data_from_file(os.path.join(args.path, args.fmap))/ (2 * math.pi)
 
 ## Dimensions check
 print('Checking dimensions...')
@@ -45,11 +48,6 @@ if len(fmap.shape) == 2:
         raise ValueError('The field map dimensions do not agree with the raw data')
 
 print('OK')
-'''print(data_in.shape)
-print(ktraj.shape)
-print(fmap.shape)
-print(args.Lx)
-print(args.grad_raster)'''
 
 ## Sequence parameters
 N = fmap.shape[0]
@@ -60,25 +58,43 @@ Nchannels = data_in.shape[-1]
 
 t_ro = Npoints * args.grad_raster
 T = np.linspace(args.TE, args.TE + t_ro, Npoints).reshape((Npoints, 1))
-seq_params = {'N': N, 'Npoints': Npoints, 'Nshots': Nshots, 't_vector': T, 't_readout': t_ro}
+if args.cart == 0:
[…]
