### 2020-10-08 #2578 @Mmj94
Thank you all so much for your patience and again thank you for all your comments and suggestions, I believe they have significantly improved the quality of this tool.

I believe I have addressed all the code comments with my last push to the `dev` branch. Please have a look at it and let me know if something else needs to be done for acceptance of the paper.

Once the checklists are completed regarding the code, I'll do a pull request to the merge branch and update the ReadTheDocs page, readme, and `PyPI` package.

There are still two open issues in the repository. I'm hoping Issue #3 will be closed once the manuscript paper is updated. To close Issue #4 I have added some helper functions to check for input dimensions while serving as documentation at the same time, please have a look at it and let me know if further action is needed.

In response to @emilljungberg specific question regarding parallel imaging data, unfortunately at this early stage, `OCTOPUS` won't be able to correct it.

I'll start addressing the manuscript issues and hopefully will get back to you soon with an update.
