### 2020-10-08 imr-framework/OCTOPUS#4 @Mmj94
Please check this commit 3e9ca4616e1f3e27580a4894755e2ee188a30849. It provides documentation regarding the data structure.
