### 2020-10-08 imr-framework/OCTOPUS#4 @emilljungberg
Thanks! Yes, it does explain the data structure a lot better. Nothing that needs to change now, but looking at the code the information in the `params` dictionary overlap with the size of the k-space trajectory, so you have the same information in two places. It could potentially be slimmed down by finding those parameters from the trajectory.
