+++ paper.md
@@ -1,62 +1,116 @@
 ---
-title: 'Off-resonance CorrecTion OPen soUrce Software (OCTOPUS)'
-tags:
-  - Python
-  - MRI
-  - off-resonance correction
-authors:
-  - name: Marina Manso Jimeno
-    orcid: 0000-0002-1141-2049
-    affiliation: 1
-  - name: John Thomas Vaughan Jr.
-    orcid: 0000-0002-6933-3757
-    affiliation: 1
-  - name: Sairam Geethanath
-    orcid: 0000-0002-3776-4114
-    affiliation: 1
+
+title: 'Off-resonance CorrecTion OPen soUrce Software (OCTOPUS)' tags:
+- Python
+- MRI
+- off-resonance correction authors:
+- name: Marina Manso Jimeno orcid: 0000-0002-1141-2049 affiliation: 1
+- name: John Thomas Vaughan Jr. orcid: 0000-0002-6933-3757 affiliation:
+  1
+- name: Sairam Geethanath orcid: 0000-0002-3776-4114 affiliation: 1
 
 affiliations:
- - name: Columbia Magnetic Resonance Research Center, Columbia University in the City of New York, USA
-   index: 1
-date: 21 July 2020
-bibliography: paper.bib
+- name: Columbia Magnetic Resonance Research Center, Columbia University
+  in the City of New York, USA index: 1 date: 21 July 2020 bibliography:
+  paper.bib
+
 ---
 
 # Summary
 
-`OCTOPUS` is a Pyton-based software for correction of off-resonance artifact in Magnetic Resoance (MR) images. It implements three different methods, it's general to the image acquisition scheme and can be run in the browser, making it a very flexible and customizable tool.
+`OCTOPUS` is a Python-based software for correction of off-resonance
+artifact in Magnetic Resonance (MR) images. It implements three different
+methods, it's general to the image acquisition scheme and can be run in
+the browser, making it a very flexible and customizable tool.
 
 # Statement of need
 
- Off-resonance is an MR artifact whose source can vary from field inhomogeneities, to differences in tissue susceptibilities and chemical shift [@Noll1991]. These phenomena can cause the phase of off-resonant (not at the resonant or Larmor frequency) spins to accumulate along the read-out direction, which can turn into blurring, geometrical distortion and degradation in the reconstructed image [@LukPat2001]. Images acquired using long readout trajectories and/or at high fields where the field homogeneity is lower, are more prone to this problem. However, such an acquisition scenario also delivers desirable propreties, such as short scanning times, gradient efficiency, motion tolerance and better signal-to-noise ratio [@Chen2008].
+Off-resonance is an MR artifact whose source can vary from field
+inhomogeneities, to differences in tissue susceptibilities and chemical
+shift [@Noll1991]. These phenomena can cause the phase of off-resonant
+(not at the resonant or Larmor frequency) spins to accumulate along the
+read-out direction, which can turn into blurring, geometrical distortion
+and degradation in the reconstructed image [@LukPat2001]. Images
+acquired using long readout trajectories and/or at high fields where the
+field homogeneity is lower, are more prone to this problem. However,
+such an acquisition scenario also delivers desirable properties, such as
+short scanning times, gradient efficiency, motion tolerance and better
+signal-to-noise ratio [@Chen2008].
 
-Multiple successful off-resonance correction methods have been reported by the literature [@Schomberg1999]. Most of them are based on Conjugate Phase Reconstruction (CPR), a method which counteracts the accumulated phase by demodulating k-space data with its conjugate [@Maeda1988]. From the original CPR, faster and more efficient implementations were developed, such as frequency-segmented CPR [@Noll1992] and Multi-Frequency Interpolation (MFI) [@Man1997].
[…]
