+++ paper.bib
@@ -113,7 +113,6 @@ title = {{Multifrequency interpolation for fast off-resonance correction}},
 volume = {37},
 year = {1997}
 }
-
 @article{Lustig2010,
 abstract = {A new approach to autocalibrating, coil-by-coil parallel imaging reconstruction, is presented. It is a generalized reconstruction framework based on self-consistency. The reconstruction problem is formulated as an optimization that yields the most consistent solution with the calibration and acquisition data. The approach is general and can accurately reconstruct images from arbitrary k-space sampling patterns. The formulation can flexibly incorporate additional image priors such as off-resonance correction and regularization terms that appear in compressed sensing. Several iterative strategies to solve the posed reconstruction problem in both image and k-space domain are presented. These are based on a projection over convex sets and conjugate gradient algorithms. Phantom and in vivo studies demonstrate efficient reconstructions from undersampled Cartesian and spiral trajectories. Reconstructions that include off-resonance correction and nonlinear ℓ1-wavelet regularization are also demonstrated. {\textcopyright} 2010 Wiley-Liss, Inc.},
 author = {Lustig, Michael and Pauly, John M.},
@@ -126,6 +125,7 @@ number = {2},
 pages = {457--471},
 pmid = {20665790},
 title = {{SPIRiT: Iterative self-consistent parallel imaging reconstruction from arbitrary k-space}},
+url = {https://people.eecs.berkeley.edu/{~}mlustig/Software.html},
 volume = {64},
 year = {2010}
 }
@@ -140,9 +140,11 @@ keywords = {Magnetic resonance fingerprinting,Off-resonance,Relaxometry,Spiral t
 mendeley-groups = {OCTOPUS-JOSS},
 pages = {63--72},
 title = {{Multi-frequency interpolation in spiral magnetic resonance fingerprinting for correction of off-resonance blurring}},
+url = {https://github.com/jostenson/MRI{\_}Ostenson{\_}MRF{\_}MFI},
 volume = {41},
 year = {2017}
 }
+
 
 @article{Jenkinson2012,
 abstract = {FSL (the FMRIB Software Library) is a comprehensive library of analysis tools for functional, structural and diffusion MRI brain imaging data, written mainly by members of the Analysis Group, FMRIB, Oxford. For this NeuroImage special issue on“20years offMRI”we have beenaskedtowrite about the history, developments and current status of FSL.We also include some descriptions of parts of FSL that are not well covered in the existing literature.We hope that some of this contentmight be of interest to users of FSL, and alsomaybe to newresearch groups considering creating, releasing and supporting newsoftware packages for brain image analysis.},
@@ -153,9 +155,11 @@ mendeley-groups = {OCTOPUS-JOSS},
 number = {2},
 pages = {782--790},
 title = {{Review FSL}},
+url = {https://fsl.fmrib.ox.ac.uk/fsl/fslwiki/FUGUE},
 volume = {62},
 year = {2012}
 }
+
 
 @article{Sutton2003,
[…]
