+++ paper.bib
@@ -421,7 +421,7 @@ doi     = {10.1038/s41586-020-2649-2}
 author = {Jyh-Miin Lin},
 title = {{Pynufft}: {Python} non-uniform fast {F}ourier transform},
 year = {2013–},
-url = “https://github.com/jyhmiinlin/pynufft”,
+url = {https://github.com/jyhmiinlin/pynufft},
 note = {Online; https://github.com/jyhmiinlin/pynufft; Dec 2016}
 }
 
