+++ paper.md
@@ -25,16 +25,15 @@ bibliography: paper.bib
 # Summary
 
 `OCTOPUS` is a Python-based software for correction of off-resonance
-artifact in Magnetic Resonance (MR) images. It implements three different
-methods, it's general to the image acquisition scheme and can be run in
-the browser, making it a very flexible and customizable tool.
+artifacts in Magnetic Resonance (MR) images. It implements three different
+methods for correction of both Cartesian and non-Cartesian data. These methods are Conjugate Phase Reconstruction (CPR), 
+frequency-segmented CPR and Multi-Frequency Interpolation(MFI).`OCTOPUS` is easy to integrate into other two-dimensional reconstruction pipelines, which makes the tool highly flexible 
+and customizable.
 
 # Statement of need
 
-Off-resonance is an MR artifact whose source can vary from field
-inhomogeneities, to differences in tissue susceptibilities and chemical
-shift [@Noll1991]. These phenomena can cause the phase of off-resonant
-(not at the resonant or Larmor frequency) spins to accumulate along the
+Off-resonance is an MR artifact which can be produced by for instance field inhomogeneities, differences in tissue 
+susceptibilities and chemical shift [@Noll1991]. These phenomena can cause the phase of off-resonant spins to accumulate along the
 read-out direction, which can turn into blurring, geometrical distortion
 and degradation in the reconstructed image [@LukPat2001]. Images
 acquired using long readout trajectories and/or at high fields where the
@@ -49,7 +48,12 @@ Phase Reconstruction (CPR), a method which counteracts the accumulated
 phase by demodulating k-space data with its conjugate [@Maeda1988]. From
 the original CPR, faster and more efficient implementations were
 developed, such as frequency-segmented CPR [@Noll1992] and
-Multi-Frequency Interpolation (MFI) [@Man1997].
+Multi-Frequency Interpolation (MFI) [@Man1997]. Frequency-segmented CPR reconstructs 
+the corrected image by combining the pixels of L base images according to each pixel value on a field map. Each base image corresponds to the data demodulated at a fixed frequency, with 
+the frequency values for each base image equally spaced within the field map frequency range.
+MFI  works in a similar way as frequency-segmented CPR, with main differences being that it 
+requires a smaller number of base images (L) and that these images are added together into the corrected image using a set of
+linear coefficients derived from the field map. 
 
 One can find limited off-resonance correction capabilities within
 existing packages. Examples are: SPIRiT [@Lustig2010], a MATLAB-based
@@ -65,27 +69,44 @@ images.
 
 All of the mentioned implementations are highly specific, defined for a
 particular k-space trajectory, application and/or include a single
-correction method. These limitations typically lead researchers to
-adjust their data in an attempt to fit it into the available pipelines
+correction method. To exemplify some: SPIRiT is devoted to correct data acquired using 
+parallel imaging methods, Ostenson's package only corrects MRF spiral data and implements 
+only one correction method and FUGUE corrects distortion solely on EPI images. These limitations typically lead researchers to
+adapt their data in an attempt to fit it into the available pipelines
 or to write their own version of the methods. Either approach results in
 a significant investment of time and effort and can generate isolated
 implementations and inconsistent results. Furthermore, most of the
-available packages are also MATLAB-based, restricting the portability,
-accessibility and customization of the code [@Ravi2018].
+available packages are also MATLAB-based, which unlike Python, requires users to pay a license fee.
 
-`OCTOPUS` - a Google Colab notebook compatible implementation - is aimed
-at filling this gap in MR off-resonance correction packages. It provides
[…]
