+++ paper.bib
@@ -439,3 +439,16 @@ url = {https://doi.org/10.7717/peerj.453},
 volume = {2},
 year = {2014}
 }
+
+@book{Gonzalez2001
+author = {Gonzalez, Rafael C. and Woods, Richard E.},
+title = {Digital Image Processing},
+year = {2001},
+isbn = {0201180758},
+publisher = {Addison-Wesley Longman Publishing Co., Inc.},
+address = {USA},
+edition = {2nd},
+abstract = {From the Publisher: Digital Image Processing has been the leading textbook in its field for more than 20 years. As was the case with the 1977 and 1987 editions by Gonzalez and Wintz, and the 1992 edition by Gonzalez and Woods, the present edition was prepared with students and instructors in mind. The material is timely, highly readable, and illustrated with numerous examples of practical significance. All mainstream areas of image processing are covered, including a totally revised introduction and discussion of image fundamentals, image enhancement in the spatial and frequency domains, restoration, color image processing, wavelets, image compression, morphology, segmentation, and image description. Coverage concludes with a discussion of the fundamentals of object recognition. Although the book is completely self-contained, a Companion Web Site (see inside front cover) provides additional support in the form of review material, answers to selected problems, laboratory project suggestions, and a score of other features. A supplementary instructor's manual is available to instructors who have adopted the book for classroom use. New Features New chapters on wavelets, image morphology, and color image processing. More than 500 new images and over 200 new line drawings and tables. A revision and update of all chapters, including topics such as segmentation by watersheds. Numerous new examples with processed images of higher resolution. A reorganization that allows the reader to get to the material on actual image processing much sooner than before. Updated image compression standards and a new section on compressionusing wavelets. A more intuitive development of traditional topics such as image transforms and image restoration. Updated bibliography.}
+}
+
+
