+++ paper.bib
@@ -440,7 +440,7 @@ volume = {2},
 year = {2014}
 }
 
-@book{Gonzalez2001
+@book{Gonzalez2001,
 author = {Gonzalez, Rafael C. and Woods, Richard E.},
 title = {Digital Image Processing},
 year = {2001},
