### 2020-10-25 3deefdf
Update documentation + minor changes
