### 2020-10-25 d7a7e07
remove source doc files with errors
