+++ doc/source/OCTOPUS.Fieldmap.rst
@@ -1,19 +1,0 @@
-OCTOPUS.fieldmap subpackage
-===========================
-
-OCTOPUS.fieldmap.simulate module
---------------------------------
-
-.. automodule:: OCTOPUS.fieldmap.simulate
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-OCTOPUS.fieldmap.unwrap module
-------------------------------
-
-.. automodule:: OCTOPUS.fieldmap.unwrap
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
