+++ doc/source/OCTOPUS.Recon.rst
@@ -1,18 +1,0 @@
-OCTOPUS.recon subpackage
-========================
-
-OCTOPUS.recon.imtransforms module
----------------------------------
-
-.. automodule:: OCTOPUS.recon.imtransforms
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-OCTOPUS.recon.rawdata\_recon module
------------------------------------
-
-.. automodule:: OCTOPUS.recon.rawdata_recon
-   :members:
-   :undoc-members:
-   :show-inheritance:
