### 2020-10-25 cccf752
Update the doc source files (lowercase)
