+++ doc/source/OCTOPUS.fieldmap.rst
@@ -1,0 +1,19 @@
+OCTOPUS.fieldmap subpackage
+===========================
+
+OCTOPUS.fieldmap.simulate module
+--------------------------------
+
+.. automodule:: OCTOPUS.fieldmap.simulate
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+OCTOPUS.fieldmap.unwrap module
+------------------------------
+
+.. automodule:: OCTOPUS.fieldmap.unwrap
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
