+++ doc/source/OCTOPUS.recon.rst
@@ -1,0 +1,18 @@
+OCTOPUS.recon subpackage
+========================
+
+OCTOPUS.recon.imtransforms module
+---------------------------------
+
+.. automodule:: OCTOPUS.recon.imtransforms
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+OCTOPUS.recon.rawdata\_recon module
+-----------------------------------
+
+.. automodule:: OCTOPUS.recon.rawdata_recon
+   :members:
+   :undoc-members:
+   :show-inheritance:
