### 2020-10-25 f5c88bd
Update requirements.txt (not compatible with newest pynufft version)
