### 2020-10-25 fce2e60
Upgrade to scipy 1.4.1 (Google Colab compatibility)
