+++ requirements.txt
@@ -5,5 +5,5 @@ opencv-python>=4.1.2.30
 pydicom>=2.0.0
 pynufft==2019.2.3
 scikit-image>=0.16.2
-scipy==1.4.0
+scipy==1.4.1
 
