### 2020-10-26 43e9708
Minor changes to manuscript + table with larger font size
