### 2020-10-27 8306bc2
add command line program to setup
