+++ setup.py
@@ -12,7 +12,7 @@ with open('requirements.txt', 'r') as f:
 
 setuptools.setup(
     name = 'MR-OCTOPUS',         # How you named your package folder (MyLib)   # Chose the same as "name"
-    version = '0.2.3',  # 0.2.1 for next stable release    # Start with a small number and increase it with every change you make
+    version = '0.2.4',  # 0.2.1 for next stable release    # Start with a small number and increase it with every change you make
     author = 'Marina Manso Jimeno',                   # Type in your name
     author_email = 'mm5290@columbia.edu',      # Type in your E-Mail
     description = 'Off-resonance correction of MR images',   # Give a short description about your library
