+++ requirements.txt
@@ -1,9 +1,45 @@
+alabaster==0.7.12
+Babel==2.8.0
+certifi==2020.6.20
+chardet==3.0.4
+colorama==0.4.4
+cycler==0.10.0
+decorator==4.4.2
+docutils==0.16
+idna==2.10
+imageio==2.9.0
+imagesize==1.2.0
+Jinja2==2.11.2
+kiwisolver==1.2.0
+MarkupSafe==1.1.1
 matplotlib==3.3.2
+MR-OCTOPUS==0.2.0
+networkx==2.5
 nibabel==3.1.1
 numpy==1.19.2
 opencv-python==4.4.0.44
+packaging==20.4
+Pillow==7.2.0
 pydicom==2.0.0
-pynufft==2019.2.3
+Pygments==2.7.2
+pynufft==2020.1.0
+pyparsing==2.4.7
+python-dateutil==2.8.1
+pytz==2020.1
+PyWavelets==1.1.1
+requests==2.24.0
 scikit-image==0.17.2
-scipy==1.4.1
-
+scipy==1.4.0
+six==1.15.0
+snowballstemmer==2.0.0
+Sphinx==3.2.1
+sphinx-rtd-theme==0.5.0
+sphinxcontrib-applehelp==1.0.2
+sphinxcontrib-devhelp==1.0.2
+sphinxcontrib-htmlhelp==1.0.3
+sphinxcontrib-jsmath==1.0.1
+sphinxcontrib-qthelp==1.0.3
+sphinxcontrib-serializinghtml==1.1.4
+tifffile==2020.9.3
+tqdm==4.50.0
+urllib3==1.25.11
