+++ requirements.txt
@@ -1,1 +1,8 @@
-scipy == 1.4.1
+matplotlib>=3.1.3
+nibabel>=3.0.2
+numpy>=1.18.1
+opencv-python>=4.1.2.30
+pydicom>=2.0.0
+pynufft==2019.2.3
+scikit-image>=0.16.2
+scipy==1.4.1
