### 2020-10-27 55ff75b
Add CLI to setup.py
