+++ setup.py
@@ -12,7 +12,7 @@ with open('requirements.txt', 'r') as f:
 
 setuptools.setup(
     name = 'MR-OCTOPUS',         # How you named your package folder (MyLib)   # Chose the same as "name"
-    version = '0.2.2',  # 0.2.1 for next stable release    # Start with a small number and increase it with every change you make
+    version = '0.2.3',  # 0.2.1 for next stable release    # Start with a small number and increase it with every change you make
     author = 'Marina Manso Jimeno',                   # Type in your name
     author_email = 'mm5290@columbia.edu',      # Type in your E-Mail
     description = 'Off-resonance correction of MR images',   # Give a short description about your library
@@ -22,7 +22,11 @@ setuptools.setup(
     packages = setuptools.find_packages(),
     include_package_data = True,
     install_requires = install_reqs,
-    scripts = ['bin/prog.py'],
+    entry_points={
+        'console_scripts': [
+            'OCTOPUS_cli = OCTOPUS.prog:main'
+        ]
+    },
     license = 'License :: OSI Approved :: GNU Affero General Public License v3',
     classifiers = [
         "Programming Language :: Python :: 3.6",
