### 2020-10-27 #2578 @Mmj94
Hi all, thank you again for your patience in this long process.
I believe we've finally implemented all of your comments in both the manuscript and code. Everything is now updated in the repository.
For the comments related to the manuscript, I've created a summary document that I'm attaching here.

Please let me know if any other action is required from my side.
Thanks again!

[Response to reviewers.pdf](https://github.com/openjournals/joss-reviews/files/5448355/Response.to.reviewers.pdf)
