### 2020-10-27 imr-framework/OCTOPUS#3 @Mmj94
Please see the new figure in the paper. You should be able to reproduce it now.
See 9ebbc28df555261d5495893cdd3fca04d838f368.
