### 2020-10-31 imr-framework/OCTOPUS#3 @emilljungberg
Thanks! Yes, I was able to reproduce the figures.
