### 2020-11-02 imr-framework/OCTOPUS#7 @emilljungberg
B0 variation in simulations
I was looking at the simulation study again and wondered about the large B0 variations you use in order to produce any visible off-resonance behaviour. In previously published studies, like Sutton et al. 2003 IEEE, they simulate their phantom experiment with off-resonance of +/- 60 Hz, which produce significant artefacts. The values you report, with the minimum variation of +/- 1600 Hz makes me wonder if there is some bug in the code, in terms of time/frequency units, or if the simulation isn't realistic?

One thing that I noticed is that the simulation is performed using, essentially, a zero echo time acquisition. The off-resonance accumulation in `numsim_cartesian.py` (and the non-cartesian example) is calculated to start from 0 and increase with the time step `dt=10E-6`. This would, to me, be equivalent to TE=0 and a readout bandwidth of +/- 50 kHz. This is a fairly high, but not unreasonable bandwidth, but the fact that the TE isn't included in the simulation might be the reason you are not seeing any significant off resonance artefacts at "sensible" B0 variations (+/- 100 Hz).

Would you consider including the TE in the simulation to make it more realistic and see if that makes the results to match the literature better?

Re: [openjournals/joss-reviews/#2578](https://github.com/openjournals/joss-reviews/issues/2578)
