### 2020-11-10 imr-framework/OCTOPUS#7 @Mmj94
I can certainly include the TE in the simulation (please note that I have included TE for the in vitro examples). However, I do not think that this inclusion will significantly change the distortion of the images.

As you pointed out, in Sutton et al. 2003 IEEE, the authors performed their simulations using a relatively homogenous field of +/- 60 Hz. Still, the resulting images presented significant off-resonance artifacts. However, the trajectory they used was a single-shot spiral with a readout time of approximately **19 ms**. The main effect of off-resonance is the accumulation of phase along the readout direction. For such a long trajectory, phase accrual is **3.8pi**, making off-resonance evident in the images. In the simulations we presented, we are using short readout trajectories, which are, in general, more clinically used and, therefore, more realistic. For this reason, we had to bump up the inhomogeneity to obtain noticeable distortion in the corrupted images, especially for the Cartesian case where readout time is only **1.28 ms**.

For demonstration purposes, I came up with a similar experiment. I designed a single-shot spiral trajectory of 19 ms read-out time and used it to sample the shepp-Logan phantom image (N=64, FOV=36cm). Off-resonance was simulated using a field map with frequency range of +/- 60 Hz. In the corrupted image (figure attached), one can see significant distortion.

![sim_singleshotspiral](https://user-images.githubusercontent.com/39202780/98731191-59ccd100-236b-11eb-87ef-00a4804813a0.png)
