### 2020-11-17 imr-framework/OCTOPUS#7 @emilljungberg
Thanks for your detailed answer! That makes sense and that simulation you show here demonstrate the effect clearly.
