### 2020-11-26 #2578 @jni
Hi everyone,

My apologies for the very long silence. It has been a very busy month and I want to thank @puolival and @emilljungberg for the thorough reviews, and @Mmj94 for your response. I need to set aside a chunk of time to consider them all, which unfortunately I can't do until the middle of next week. Apologies that I did not communicate this sooner — I kept thinking that I could do something "in the next few days" and I never found the time. Next week a big deadline passes and I really hope to be able to set some time for this.

Thank you all again. 🙏
