### 2020-12-17 #2578 @emilljungberg
Hi,

First of all, thanks for your work on the revisions to the first round of reviews @Mmj94. I think it has really strengthened the manuscript and the toolbox. After looking through the revised manuscript and the repository I have a few points I'd like to highlight though:

In its current state, I think the statements in the paper about OCTOPUS reach beyond its current capacity. The summary reads

> OCTOPUS is easy to integrate into other two-dimensional reconstruction pipelines, which makes the tool highly flexible and customizable.

With only being able to use 2D and single-coil data, that is quite a big limitation to me. I think this then comes back in the "Statement of need" section where it is compared to other packages. The paragraph begins with

> One can find limited off-resonance correction capabilities within existing packages.

I think _limited_ is the wrong word here, _optimised_ would be better suited as they are developed to be optimised for a certain application, which is understandable as pulse sequences are very different and can benefit from making certain assumptions.  Furthermore, I don't think the assessment of the current methods, later on, is fair.

> SPIRiT is devoted to correct data acquired using parallel imaging methods, Ostenson’s package only corrects MRF spiral data and implements only one correction method and FUGUE corrects distortion solely on EPI images

Basically all MRI scans today are acquired using multi-channel coils, which would make SPIRiT (to my understanding) applicable to almost all data available out there. MRF scans are highly undersampled acquisitions which understandably would require special considerations. Unless OCTOPUS can be shown to handle all the above situations I think it would be fairer to consider these packages as being optimised rather than limited.

The main point is that the final section then states that OCTOPUS is going to fill this gap in the literature, which to the reader then would imply that it can deal with any acquisition method, thus being able to deal with the type of data for which multiple specialised packages have been developed, as well as data that isn't covered by these specialised methods. While I do think this is a great initiative and I really appreciate accessible software written in Python which we all can learn from, OCTOPUS does not fill this gap in the current state I'm afraid. In essence, my concerns are highlighted in the limitations section which to me are too big to not be implemented in the first version, which includes

1. 3D data
2. Multiple receiver coils (Not necessarily being able to deal with aliased data from parallel imaging, but demonstrating that it can deal with multiple receiver coils with different sensitivity profiles).

I'd consider both of these aspects key to making it flexible and practically useable in many situations. 3D imaging is becoming more popular, and multi-channel coils are basically standard today. You mention in the reviewer comments that you see this as a version 0 of the software and that development will be ongoing. Therefore I would suggest further development of the toolbox as planned and then release it in JOSS, in what then might be version 1.0. To my understanding, this wouldn't require a new review process, we can keep this thread going but extend the review process basically.

I also think delaying publication until a more developed version would be a benefit for your team and the software, as you would then showcase a more comprehensive package that will be more applicable to users demands, thus reaching a wider audience.

This is a really great project and I thank you for all your work on it. With further development including 3D data and multi-coil support, which I understand might already be ongoing, I think this will be a really useful tool for researchers and students, and also to learn more about how off-resonance correction in MRI works.
