### 2020-12-17 #2578 @Mmj94
Dear reviewers,

First of all, thank you, @emilljungberg, for your comments and inputs. I totally respect your opinion, but I would like to clarify the multi-coil support limitation you're referring to.
`OCTOPUS` would be able to correct multi-coil data (as long as it's not undersampled) using the **current** version of the tool. The user could correct each channel data independently and then use a channel combination method of his/her choice. In fact, this is what the example code `ORC_main.py` does, combining channels using sums of squares after correction.
Sorry if the manuscript does not convey that feature correctly and please, let me know if you have further doubts regarding this point.
