### 2020-12-17 #2578 @emilljungberg
Thanks @Mmj94 for your rapid reply. Sorry that I missed that reg multiple-coils, now that you mention it I remember commenting on the sum of square coil combination in my first round of comments, my bad. Maybe you could clarify this in the Functionality and Limitations section, that it works with multi-coil data but not parallel imaging acceleration. Out of curiosity, would it work to apply this to coil-combined complex-valued data?

You can consider that point addressed then. Thanks!
