### 2020-12-17 #2578 @Mmj94
Thank you.

Regarding your question, I usually correct each channel independently and then combine them since that's what I've read is most common in the literature. But yes, the tool would work on coil-combined first complex data as well.
