### 2020-12-18 #2578 @jni
Hi @Mmj94,

@emilljungberg and I had a private conversation about this, as he felt pretty strongly about the above requirements pre-publication, but wasn't sure about JOSS's editorial policy regarding significant revisions. I'm sorry that things were taking so long. Emil essentially summarised it above — that adding ~~two~~ one feature~~s~~ would significantly strengthen the software's position and utility to the community, and thus that publication before that feature is added would be premature. I'm glad that one of Emil's major points is already addressed! 🎉

Generally, my position is that major revisions are a pain in the neck and should not be required for publication — a certain amount of work has been done and the paper reports on that work. But, in this case, I checked with others at the JOSS editorial team and there was consensus that 3D support should be added before publication, given the increasing prevalence of 3D acquisition.

Do you think the 3D data acquisition is something that you could address before publication? On our end at JOSS we'd be happy to wait, as we think it would increase the chance of reuse and community adoption, which is one of the goals we want to pursue here at JOSS.

I'm sorry about this — I have been on the other end of the "major revisions" stick and it sucks and it pains me to write this. But I hope you will see that this would at a minimum be a really cool feature that will mean that lots more people will find your software useful, so that the work is worth doing anyway.

One question I would have about the whole process, though, is whether you in fact have easy access to 3D acquisition data.

Oh and one last thing: one of the JOSS editors flagged that the name Octopus is already used by a [tomographic reconstruction framework](https://iopscience.iop.org/article/10.1088/0957-0233/15/7/020), which could be confusing. Software name clashes are nothing new and this is not a problem from our end, but something to consider since OCTOPUS is at an early stage and changing the name might not be too traumatic at this point. (Though you do have a cute logo. 😊)

Please don't hesitate to reach out with further questions or concerns about the above.
