### 2021-01-10 #2578 @Mmj94
Hi @jni and @emilljungberg,

First o all, sorry for my delayed response (winter break + deadlines).

We believe that we could do 3D data acquisition if that's what you require for publication.
However, could you be a bit more specific about it?

We wondered if you mean 3D multi-slice acquisition, which is already supported, or you mean 3D k-space trajectory acquisition. In the case of the latter, could you provide an example trajectory? We have already used `OCTOPUS` to correct images from a 3D stack-of-spirals sequence, for example.

Note that for applications where off-resonance is an issue, one would likely use the shortest readout possible for a given trajectory, as off-resonance artifacts worsen with longer durations. For this reason, I don't know if using a 3D k-space trajectory that usually entails longer acquisition times would be relevant. Of course, I'm open to discussion.
