### 2021-01-14 #2578 @emilljungberg
Hi again everyone,

Regarding 3D coverage, I was mainly thinking that it would be good to clarify how that could be achieved as it would open up your package to more researchers. If you have done 3D stack of spirals, that's great and shows that it can be done. Any chance you can include an example of those results in the manuscript? So if this capability exists, then maybe you can reword it to something like "Correctio of 3D volumes is possible by processing individual 2D slices (if possible)", as in the case of stack-of-stars/spirals.

Regarding using short readout durations for minimised off-resonance artefacts, I agree with you, shorter readouts are necessary to reduce off-resonance artefacts. But if you reduce the readout for that purpose, you are unlikely to require any corrections. Isn't it in cases where you have long readouts (EPI, spirals) and thus strong artefacts that you actually need these corrections? If your simulation sequence requires B0 variations on the order of kHz, that suggests that such sequence never would require off-resonance correction. And thus, I'm not sure if the simulations you include in the paper are particularly helpful for the average reader.

I appreciate that you had to increase the off-resonance range to see an effect with your short Cartesian readout but that range is way beyond what you ever would observe in an MRI scanner, and thus not a useful reference for someone looking at implementing this for their own use. I understand that it clearly demonstrates that the methods you implemented works as expected, but I think it would be more useful to include a more realistic example. For example, the papers you reference have off-resonance ranges between:

- 120 Hz (Fessler et al. 2005)
- 65 Hz (Luk-Pat et al. 2001)
- 100 Hz (Man et al. 1997)

Would you consider changing the sequence used for your simulation to one that would be strongly affected by off-resonance variations on the order of +/- 200 Hz or something like that? Also, if you do have these large B0 variations in the image, I also think you would start to see intra-voxel dephasing due to B0 variation within a single voxel and from that major signal dropout.

If you do want to keep the simulations as is, I would strongly recommend adding in a comment about the large off-resonance variations.
