### 2021-01-22 #2578 @Mmj94
Hi @emilljungberg,

This is an update just to let you know that I'm working on implementing the changes you proposed to the paper's simulations.
Specifically:

1. **Single-shot spiral and EPI** trajectories for off-resonance ranges of &pm;100, 150, and 200 Hz.
2. Try including some results from correction of **stack-of-spirals or 3D multi-slice spiral** sequences. Note that the results I mentioned earlier about the stack-of-spirals are fMRI images with other types of artifacts apart from off-resonance. For this reason, I don't think I should include those in the manuscript. I'm trying to retrieve some phantom data acquired using that sequence, or there's the possibility I'll have to acquire such data (this may take some extra time).

Would these two items change your mind about the publication of the tool? Of course, I'll refine the manuscript with the changes you suggested too.

Finally, I would like to clarify what I meant with  "For this reason, I don't know if using a 3D k-space trajectory that usually entails longer acquisition times would be relevant". I referred to three-dimensional trajectories in k-space, such as 3D cones, which would require excessively long readout durations.

Thank you,
