### 2021-02-08 cd3a56e
Paper changes (modify simulation figure and add SoS correction)
