+++ paper.bib
@@ -440,6 +440,19 @@ volume = {2},
 year = {2014}
 }
 
+@article{Shepp1974,
+abstract = {The authors generalize a reconstruction algorithm proposed by Ramachandran. The modified weighting function simultaneously achieves accuracy, simplicity, low computation time, as well as low sensitivity to noise. Using a simulated phantom, the authors compare the Fourier algorithm and a search algorithm, very similar to one described by G. N. Hounsfield. The search algorithm required 12 iterations to obtain a reconstruction of accuracy and resolution comparable to that of the Fourier reconstruction, and was more sensitive to noise. To speed the search algorithm by using fewer iterations leaves decreased resolution in the region just inside the skull which could mask a subdural hematoma.},
+author = {Shepp, L. A. and Logan, Benjamin F.},
+doi = {10.1109/tns.1974.6499235},
+issn = {00189499},
+journal = {IEEE Transactions on Nuclear Science},
+mendeley-groups = {OCTOPUS-JOSS},
+pages = {21--43},
+title = {{FOURIER RECONSTRUCTION OF A HEAD SECTION.}},
+volume = {21},
+year = {1974}
+}
+
 @book{Gonzalez2001,
 author = {Gonzalez, Rafael C. and Woods, Richard E.},
 title = {Digital Image Processing},
