+++ paper.md
@@ -29,12 +29,12 @@ bibliography: paper.bib
 `OCTOPUS` is a Python-based software for correction of off-resonance
 artifacts in Magnetic Resonance (MR) images. It implements three different
 methods for correction of both Cartesian and non-Cartesian data: Conjugate Phase Reconstruction (CPR), 
-frequency-segmented CPR and Multi-Frequency Interpolation(MFI). `OCTOPUS` is easy to integrate into other two-dimensional reconstruction pipelines, which makes the tool highly flexible 
+frequency-segmented CPR and Multi-Frequency Interpolation(MFI). `OCTOPUS` is easy to integrate into other two and three-dimensional reconstruction pipelines, which makes the tool highly flexible 
 and customizable.
 
 # Statement of need
 
-Off-resonance is an MR artifact which can be produced by for instance field inhomogeneities, differences in tissue 
+Off-resonance is an MR artifact which occurs due to field inhomogeneities, differences in tissue 
 susceptibilities and chemical shift [@Noll1991]. These phenomena can cause the phase of off-resonant spins to accumulate along the
 read-out direction, which can turn into blurring, geometrical distortion
 and degradation in the reconstructed image [@LukPat2001]. Images
@@ -51,7 +51,7 @@ phase by demodulating k-space data with its conjugate [@Maeda1988]. From
 the original CPR, faster and more efficient implementations were
 developed, such as frequency-segmented CPR [@Noll1992] and
 Multi-Frequency Interpolation (MFI) [@Man1997]. Frequency-segmented CPR reconstructs 
-the corrected image by combining the pixels of L base images according to each pixel value on a field map. Each base image corresponds to the data demodulated at a fixed frequency, with 
+the corrected image by combining the pixels of "L" base images according to each pixel value on a field map. Each base image corresponds to the data demodulated at a fixed frequency, with 
 the frequency values for each base image equally spaced within the field map frequency range.
 MFI  works in a similar way as frequency-segmented CPR, with main differences being that it 
 requires a smaller number of base images (L) and that these images are added together into the corrected image using a set of
@@ -96,18 +96,18 @@ NiBabel[@Nibabel], Matplotlib [@Matplotlib], OpenCV [@itseez2015opencv], Pydicom
 The expected output is an image with recovered, sharper edges and undistorted shape.
 
 Also, `OCTOPUS` corrects off-resonance independently of the trajectory used to acquire the data being Cartesian or non-Cartesian. 
-The input of the correction methods could be either image or raw data, although using raw data is more efficient
-and may avoid trajectory-dependent artifacts in the case of non-cartesian data. `OCTOPUS` is also able to correct 3D multi-slice and multi-channel data by feeding it to the tool in a slice and channel-wise manner and then applying channel combination with the user's method of choice.
+The input of the correction methods could be either image or raw data. However, using raw data as input is more efficient
+and may avoid non Cartesian trajectory-dependent artifacts. `OCTOPUS` is also able to correct 3D multi-slice and multi-channel data by feeding it to the tool in a slice and channel-wise manner and then applying channel combination with the user's method of choice.
 
-At the current stage, the software limitations include: correction restricted to data two-dimensionally acquired in the absence of 
+Presently, the software limitations include: correction restricted to data acquired in the absence of 
 acceleration techniques; long correction times for large datasets and degraded correction quality in the presence of highly-inhomogeneous
 fields. Additionally, the tool has been only tested on Cartesian, EPI and spiral data.
 
 # Short demo
 To illustrate the usage of the package we performed in silico numerical
 simulations using a single-shot EPI trajectory, a single-shot spiral trajectory and a
[…]
