+++ paper.md
@@ -124,11 +124,11 @@ In both experiments, 'OCTOPUS' has successfully corrected the
 off-resonance induced blurring and/or geometrical distortion. Note how the EPI corrupted images show geometric distortion in the phase-encode direction while spiral corrupted images show blurred and distorted edges.
 
 To test the effect of noise on the correction performance we introduced different levels of noise to a Cartesian trajectory-based simulation and measured the peak signal-to-noise ratio (pSNR) and Structural Similarity Index (SSIM). We increased the off-resonance range of the field maps for this experiment to observe significant distortion in the Cartesian images, which are not as prone to off-resonance due to their extremely short readout periods.
-![Effect of different noise leves on OCTOPUS correction performance measured using pSNR and SSIM.](JOSS_figs/noise_sim.png)
+![Figure 2: Effect of different noise leves on OCTOPUS correction performance measured using pSNR and SSIM.](JOSS_figs/noise_sim.png)
 As expected, PSNR and SSIM are reduced as the off-resonance range widens and the noise level in the original image increases. Nevertheless, in all cases, the three implemented methods improve the metrics with respect to the off-resonance corrupted image.
 
 Finally, to demonstrate the correction capabilities in 3D multi-slice and multi-channel data, we corrected phantom images of a Stack-of-Spirals acquisition with matrix size of 72x72, FOV=240 mm<sup>2</sup> and 54 slices. The images were acquired on a Siemens 3T Prisma scanner using a 20-channel head coil. Figure 3 shows three representative slices and their off-resonance corrected versions. The regions of the images highlighted in red show improved image quality and enhaced edges.
-![Off-resonance correction of three slices of a Stack-of-Spirals 3D acquisition.](JOSS_figs/SoS_ORC.png)
+![Figure 3: Off-resonance correction of three slices of a Stack-of-Spirals 3D acquisition.](JOSS_figs/SoS_ORC.png)
 
 # Acknowledgements
 
