### 2021-02-09 09a4283
fix labels for figs 2&3
