### 2021-02-09 7de7dc5
Fix figure labels
