### 2021-02-09 #2578 @Mmj94
Hello all,

I'm back with the latest updates on the manuscript. I hope I have assessed all you comments @emilljungberg.

_If you have done 3D stack of spirals, that's great and shows that it can be done. Any chance you can include an example of those results in the manuscript? So if this capability exists, then maybe you can reword it to something like "Correctio of 3D volumes is possible by processing individual 2D slices (if possible)"_

- Regarding the stack-of-spirals correction, I have included the correction results of three slices in Figure 3.
- I've also mentioned that 3D volumes can be processed in a slice and channel-wise manner (Lines 73-75)

_Would you consider changing the sequence used for your simulation to one that would be strongly affected by off-resonance variations on the order of +/- 200 Hz or something like that?_

- I've followed your advice and I've modified Figure 1 to show the effects of field homogeneity in the ranges -/+100, -/+150 and -/+200 Hz. The simulations use single-shot EPI and spiral trajectories, which are extremely susceptible to even small field inhomogeneities due to their long readout periods.

Please let me know if you have further inputs/questions.

Sorry for all the PDFs generated above, please refer to the last one.

PS: I'm having some trouble including both my last names without using a hyphen. @jni any advice with this issue?
