### 2021-02-10 #2578 @jni
> PS: I'm having some trouble including both my last names without using a hyphen. @jni any advice with this issue?

Hmm, a very good question! I don't see how the current authorship format can disambiguate between first, middle, and last names. @openjournals/joss-editors any advice on this?
