### 2021-02-10 #2578 @jni
I've been scouring the  for yaml-style authorship definitions, and various pandoc doc pages, and I wonder whether LaTeX formatting will work here: `- name: Manso Jimeno, Marina`?

I think I've tracked down the bit of the whedon source code that does the name parsing, but my Ruby is not good enough to follow the parsing logic:

https://github.com/openjournals/whedon/blob/92346a0773a4003bf0ef361b661dc2787f492936/lib/whedon/author.rb#L2

Which appears to use this class:

https://github.com/chorn/nameable/blob/4dc46df3ee3080a054eb639839d39a66c8568259/lib/nameable/latin.rb#L5

It does *not* seem to me that the Lastname, Firstname pattern will work, but I wonder whether quotes will work somehow. I think I'll just ping @arfon for this one, since he wrote that `author.rb` code! 😬
