### 2021-02-10 #2578 @arfon
> Hmm, a very good question! I don't see how the current authorship format can disambiguate between first, middle, and last names. @openjournals/joss-editors any advice on this?

Unfortunately I will have to fix this in post-production (i.e., once the paper is already accepted and published). In the longer-term, we're planning on updating the metadata that JOSS allows for author names (to make it explicit what the last/family name is) but we can't do that yet sorry.
