### 2021-02-10 #2578 @emilljungberg
Hi @Mmj94

Thanks for sending the revised version of the manuscript. I think these revisions looks really nice. The EPI and spiral readout off-resonance corrections looks great and do show the type of artefacts that are expected at the given off-resonance range. One question though, why did you change the B0 map model from the last version? The previous version had like a smooth almost like linear field map.

In the second part of the results where you do the noise comparison, you use the Cartesian acquisition which is no longer defined in the latest revision. Would you consider doing that analysis with the EPI or spiral results instead to make it consistent?

It is great to see the phantom results included in this version. On the top row, I'm not sure what improvement I'm expected to see, can you clarify? In the bottom row I can see that the three diagonal elements are better resolved after correction, would you consider making an arrow to them instead of the rectangle? It would help to both know what to focus on and also to see the edges of the black box which are currently covered by the rectangle.

Otherwise, I have no further comments. Looks great!

Thanks!
