### 2021-02-18 f3b18c5
Changes to noise simulation and SoS correction figures
