### 2021-02-18 #2578 @Mmj94
Hi @emilljungberg,

I updated the manuscript with your latest comments:

1. Modified the noise simulation experiment (Figure 2) from Cartesian to EPI trajectory to make it consistent.
2. Improved regions on the images after correction (Figure 3) are now indicated with an arrow.

As per your questions:

>  why did you change the B0 map model from the last version?

I believe this field map is more "realistic" and resembles more a brain field map.

> It is great to see the phantom results included in this version. On the top row, I'm not sure what improvement I'm expected to see, can you clarify?

Inside the inner circular structure of the phantom, one can appreciate smaller circles. I believe the edges of these circles are more discernable after off-resonance correction.

Thank you again!
