### 2021-02-19 #2578 @emilljungberg
Hi @Mmj94

The PDF looks nice, I think I see the changes in the phantom data without the circles so that's greatly appreciated. I don't have any further comments on the manuscript then.

However, after these changes to the examples I downloaded the code again and tried to replicate the results from the paper but I'm afraid I'm not able to. When I run
```sh
python3 numsim_epi.py
```
It prints a field map that looks like this, which isn't the same as in the manuscript
![Fieldmap_EPI_75Hz](https://user-images.githubusercontent.com/3906827/108523229-93e03400-72c5-11eb-846b-4b247d21f522.png)

And this is the final result from the EPI simulation
![EPI_results](https://user-images.githubusercontent.com/3906827/108523708-24b70f80-72c6-11eb-9207-039211158113.png)

Is there a chance you didn't push the latest changes to the repository? I notice the the off resonance range is different and also the shape of the fieldmap as shown above.
