### 2021-02-22 8598b0a
reproduce figures in numerical simulations
