+++ OCTOPUS/Examples/numsim_spiral.py
@@ -11,6 +11,7 @@ import matplotlib.pyplot as plt
 from skimage.data import shepp_logan_phantom
 from skimage.transform import resize
 from skimage.segmentation import flood_fill
+from skimage.filters import gaussian
 
 import OCTOPUS.ORC as ORC
 import OCTOPUS.fieldmap.simulate as fieldmap_sim
@@ -21,8 +22,8 @@ from OCTOPUS.recon.rawdata_recon import mask_by_threshold
 # Original image: Shep-Logan Phantom
 ##
 def numsim_spiral():
-    N = 192  # ph.shape[0]
-    ph = resize(shepp_logan_phantom(), (N, N)).astype(complex)
+    N = 128  # ph.shape[0]
+    ph = resize(shepp_logan_phantom(), (N, N))#.astype(complex)
     plt.imshow(np.abs(ph), cmap='gray')
     plt.title('Original phantom')
     plt.axis('off')
@@ -39,9 +40,8 @@ def numsim_spiral():
     ##
     # Spiral k-space trajectory
     ##
-    dt = 10e-6
-    # ktraj = np.load('sample_data/ktraj_sutton2003.npy')
-    ktraj = np.load('sample_data/ktraj.npy') # k-space trajectory
+    dt = 4e-6
+    ktraj = np.load('sample_data/SS_sprial_ktraj.npy') # k-space trajectory
 
     plt.plot(ktraj.real,ktraj.imag)
     plt.title('Spiral trajectory')
@@ -55,7 +55,7 @@ def numsim_spiral():
     ##
     # Simulated field map
     ##
-    fmax_v = [250, 500, 750] # Hz
+    fmax_v = [100, 150, 200]  # Hz
 
     i = 0
     or_corrupted = np.zeros((N, N, len(fmax_v)), dtype='complex')
@@ -64,12 +64,10 @@ def numsim_spiral():
     or_corrected_MFI = np.zeros((N, N, len(fmax_v)), dtype='complex')
     for fmax  in fmax_v:
 
-        field_map = fieldmap_sim.realistic(np.abs(ph), mask, fmax)
-        ### For reproducibility
-        # dst = np.zeros((N, N))
-        # field_map = cv2.normalize(np.load('M2.npy'), dst, -fmax, fmax, cv2.NORM_MINMAX)
-        # field_map = field_map * np.load('mask.npy')
-        # field_map = fieldmap_gen.fieldmap_bin(field_map,5)
+        SL_smooth = gaussian(ph, sigma=3)
+        field_map = cv2.normalize(SL_smooth, None, -fmax, fmax, cv2.NORM_MINMAX)
+        field_map = np.round(field_map * mask)
+        field_map[np.where(field_map == -0.0)] = 0
         ###
         plt.imshow(field_map, cmap='gray')
         plt.title('Field Map +/-' + str(fmax) + ' Hz')
@@ -80,7 +78,7 @@ def numsim_spiral():
     ##
     # Corrupted images
     ##
-        or_corrupted[:,:,i], _ = ORC.add_or_CPR(ph, ktraj, field_map, 1, seq_params)
+        or_corrupted[:,:,i], ksp_corrupted = ORC.add_or_CPR(ph, ktraj, field_map, 1, seq_params)
         corrupt = (np.abs(or_corrupted[:,:,i]) - np.abs(or_corrupted[...,i]).min())/(np.abs(or_corrupted[:,:,i]).max() - np.abs(or_corrupted[...,i]).min())
         #plt.imshow(np.abs(or_corrupted[:,:,i]),cmap='gray')
         plt.imshow(corrupt, cmap='gray')
[…]
