### 2021-02-22 #2578 @Mmj94
Hi @emilljungberg,

I thought about these example scripts as templates that the user could modify to experiment with the effect of different field maps, off-resonance ranges, k-space trajectories, etc. However, I just push the code so the default parameters can reproduce the figures on the paper.

Hope this helps!
