### 2021-02-24 #2578 @emilljungberg
Hi @Mmj94

Thanks! I did notice now that the reference to running the scripts is not in the paper anymore which it was originally, I think that's why I was confused about not being able to reproduce the results. Having the example scripts reflect the figures in the paper is nice as it gives the user a direct check that they got it all working correctly.

I ran the `numsim_epi.py` and I get the same figures as you got so that is all fine.

With that, I don't have any more comments. Nice work!
