### 2021-02-25 #2578 @jni
@Mmj94 here we go!

Next steps: could you make a new tagged release of the software, archive it (e.g. with Zenodo), then report the version number and DOI here?
