### 2021-03-02 #2578 @Mmj94
Version number: `v1.0.0`
DOI: 10.5281/zenodo.4574444
