### 2021-03-04 #2578 @jni
@arfon as a reminder, we need to remove the hyphen from Marina's last name in the final pdf and paper metadata. :pray:

Thank you so much everyone for your patience with this review, thank you @puolival and @emilljungberg for reviewing, and thank you @Mmj94 for all the iteration! :tada:
