### 2021-03-04 #2578 @danielskatz
@Mmj94 - please see the changes suggested in https://github.com/imr-framework/OCTOPUS/pull/8 and merge them, or let me know what you disagree with, then we can proceed to final acceptance and publication
