### 2021-03-04 #2578 @arfon
> 👋 @arfon - given the specific handling needed, can you do the final accept here? (rather than me doing it and then you fixing it afterwards)

Certainly. I've gone ahead and processed this manually.

@puolival, @emilljungberg - many thanks for your reviews here and to @jni for editing this submission! JOSS relies upon the volunteer efforts of people like you, and we simply wouldn't be able to do this without you ✨
